requirements.txt
Django
psycopg2-binary
python-decouple
